Skip endonym in linguist settings list when not available
The settings dialog iterates all members of the QLocale::Language
enum, producing an entry for each. It tried to add the endonym of each
language to its entry; however, QLocale(language) gets the default
locale if there is no CLDR data for the given language; and the
default locale's endonym is no use to us. So don't attempt to include
the endonym of a language unless the locale from which we get it
actually does use the requested language. Also mention why English
skips its endonym.

Pick-to: 6.3 6.2 5.15
Fixes: QTBUG-102832
Change-Id: Ia4583727beb247e4775f443d98089db3f02a17d2
Reviewed-by: Kai Koehne <kai.koehne@qt.io>