QDoc: Add support for generating POSIX and DOS paths in catch_generator
To support the testing of QDoc parts that relates to interfacing with
the filesystem, a series of functions in "path_generator.hpp", that
specialize `path` to support a format that is suitable for Linux,
Windows and macOS were added.

The new functions allow the generation of paths that approximately
respect the POSIX and tradition DOS formats, that are, respectively,
suitable for Linux/macOS and Windows.

Some special care was taken into easing the use of the generated paths
for testing, such that not all possible paths are able to be generated.
The form of paths that are removed are general considered less
important, as they are not expected to be a common case in QDoc,
than the ease of testing that the removal provides.

To support the construction components for those paths, some generating
function were added in `qchar_generator.hpp` to allow an easier access
to some common character sets.

Two new namespaces were added to `namespaces.hpp` to enclose this newly
added code, as it didn't fit completely in the previously available
namespaces.

To support the general use of `OneOfGenerator` functions in an
expression context, considering that it requires a vector of move-only
types, a new utility was added.

The utility lives in `move_into_vector.hpp`, under `utilities/semantics`
and provides a single function that simplifies the construction of a
vector of move-only types.

Partial tests were added for the new functions in `path_generator.hpp`.
Some required tests were not added right away as they require some
thought on the testing strategy.

Some tests for the new functions in `qchar_generator.hpp` were added.

Task-number: QTBUG-100381
Change-Id: I687231791a1e67e6c2821a6584bab25940222e54
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>