QDoc: Add tests for FileResolver and related types
A new directory, `qdoc`, was added under the auto tests directory for
QDoc.

The directory contains a test project that is intended to hold all
testing files for QDoc itself.

The tests will follow the usual Catch2 file formats, and contains a main
file holding the required boilerplate and a CmakeLists.txt file that
defines the test executable.

The directory structure under the new `qdoc` directory will resemble, as
much as possible, the source directory for QDoc, simplifying the
positioning of test files and improving the searchability of the
directory when working on QDoc source files.

Based on the infrastructure provided by `catch_generators`, a series of
test files testing some of the more obvious definition for
`FileResolver`, `DirectoryPath` and `FilePath` were added.

The `CmakeLists.txt` file under the auto tests directory for QDoc was
modified to include the newly added project.

Task-number: QTBUG-100381
Change-Id: I62486b887ded084bafd5006df1ff052894b260ae
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>