QDoc: Avoid skipping the first value in OneOfGenerator and PathGenerator
Catch2 requires that a call to `get` on a generator that was just
constructed produces a correct value, that is, the generator must be
non-empty and its first value must be produced on construction.

Generally, this simply means that a call to `next` is executed in the
constructor of a generator to produce the first value.

When a generator implementation wraps one or more other generators, this
is generally not possible. As `next` would obviously need to advance the
wrapped generators, losing the first value that the generators contain.

There are different ways to solve this issue.

For the generators in `catch_generators`, this generally meant that a
small amount of code duplication between the constructor and the
implementation of `next` was used, as a simple enough way to retain the
first value that the generator uses.

Apart from not scaling easily to more complex generators, this was
error-prone.
Indeed both PathGenerator and OneOfGenerator incorrectly implemented
this such that values of some of the wrapped generators would be skipped
based on the randomized elements that were produced.

A more complex solution could be attained to solve the problem,
generally requiring a small cache and checked access.

But this solution, similar to to other solutions like sharing the
implementation with `next` trough abstraction, is generally complex and
semi ad-hoc depending on the generator.

To help avoiding similar issues in the future, and to solve the current
bugged implementation, a small wrapper for generators,
`GeneratorHandler`, was added.

Wrapping a generator with a `GeneratorHandler`, makes it so the
semantics of the generators changes such that the first call to `get`
should be preceded by a call to `next`.

This abstracts away the difference between accessing the first element
and the subsequent elements of a generator, allowing the use of the
obvious `next` implementation without skipping the first value of the
underlying generators.

The new generator was added under `utilities/semantics`, in the new
header `generator_handler.hpp`, and a small test suite was added for it.

The `CmakeLists.txt` file for `tests` was modified to include the new
suite.

Contrary to our normal hierarchy for the project, the generator was not
added under `generators`.
This is because this generator, while implementing the generator
interface, does not actually respect the semantic that is
expected of a generator.

While a generator that produced the same benefits while respecting the
semantics of Catch's generators was possible, indeed one such
generalization of `GeneratorHandler` would be a generator that
duplicates the first element of a wrapped generator, it was specifically
chosen to use an incorrect generator to emphasize its role as an
implementation detail with a specific scope.

This further allowed us to avoid requiring a specific guarantee for what
would happen if a call to `get` was issued before a call to `next` on a
`GeneratorHandler`, allowing us, instead, to use an assertion to catch
those kind of uses of the generator, albeit they may work based on the
current implementation.

Both the generators that were affected by a first-value access bug and
those that weren't were modified to use `GeneratorHandler`.
An exception was provided for `CycleGenerator`, whose implementation is
better served by specifically obtaining the first element of the wrapped
generator in the constructor.

A test to ensure that the bug was resolved was added to the touched
generators.

Task-number: QTBUG-100381
Change-Id: I1f70015901b4451c221328d90195535585bb50da
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>