QDoc: Require the directory to be executable for DirectoryPath
Up to now, `DirectoryPath`, similar to `FilePath`, required the pointed
to directory to exist and be readable.

This has shown to be an insufficient condition for paths, at least on a
Linux system, that end in "." or "..".

In particular, consider the path P pointing to a directory that is
readable but not executable.

Using `QFileInfo`/`QDir` with P, will correctly identify the path as
pointing to directory that exist and is readable.

If a "/." is appended to P, using `QFileInfo`/`QDir` will fail to identify
the directory as existing and readable.

This is because, at least on Linux, "." is an (almost)hard-link on the
filesystem and, being an actual child of the directory pointed to by P,
requires execution permission on its parent directory for
`QFileInfo`/`QDir` to `stat` it to pass the `exists`/`isDir` tests.

Similar issues comes from path ending in `..`.

We seem to have paths that use this form in the current documentation,
such that it would be unwise to not support them.
Furthermore, all those paths seem to refer to directories that have the
execution permission that are required to read those problematic paths.

To solve the issue then, we require `DirectoryPath` to guarantee
execution permission too, so those paths can be supported correctly.

This generally add a non-necessary condition for most other paths, but
we currently expect this to not be a problem, albeit this may change in
the future.

The tests for `DirectoryPath` were changed accordingly.

Some small changes were added to the tests for `DirectoryPath` and
`FilePath` to ensure that the used `QTemporaryDir`s are able to be
correctly cleaned when the tests end.

Task-number: QTBUG-100381
Change-Id: I475ca03e8bd94bd8b0eb0732b87e07e015abfb3f
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>