Update commercial license headers
Updated header.COMM to the files in tqtc-qttools.
Examples, tests, or documentation files are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a
change in the line count causes failures in tests.

Task-number: QTQAINFRA-4938
Change-Id: Ic6c4ccfb1b31de7377f7e43aef632cbc7cd22f02
Reviewed-by: Kai Koehne <kai.koehne@qt.io>