QDoc: Save QML methods return type/parameters in index files
QDoc saves information about the elements of a specific documentation
project inside an xml-formatted index file that are later used by
dependent projects to provide information about or linking to elements
of the dependent upon project.

Previously, QDoc would save basic information about QML methods but
would not save any information about the return type or
parameters of the same.

This generated some discrepancies between the information available to
the project that contains those elements and the one provided by
projects that depends on those elements.

For example, it was impossible for `\sincelist` commands in `qtdoc` to
show return type information for QML methods in `qtquick3d`.

To avoid those discrepancies, the return type and parameters information
for QML methods, or more generally non CPP nodes, are now saved in the
index files.

The changes to the code specifically tried to preserve the format and
ordering of internal parameters for the index files as much as possible,
albeit this might sligthly impact the readability of the code.

As a side effect of the change some of the all-members pages for QML
related classes now show the full information for their members, making
them consistent with the all members pages for CPP classes.

Fixes: QTBUG-103368
Change-Id: I664d7c4cb4df1e11372b11ae3537068c0c6fa9db
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
(cherry picked from commit f0f9c1d99a360ead849ab2c199c9801b80fff6a9)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>