QDoc: Contextualize QML methods/properties in sincelists
QDoc allows the construction of a summary page based on user-provided
`\since` information through the `\sincelist` command, to present an
overview of the changes occurring between versions of the software.

The list presents, among other, newly introduced QML methods and
properties, but does so in a flat-way where the methods/properties
not contextualized to the QML types holding them.

To provide a more easily digestible and useful list of the new QML
features, the presentation of QML methods and properties was modified
follow the same presentation, for consistency, that CPP's member
functions follow, showing each new element under a reference to the
holding parent.

Furthermore, the presentation of QML properties was modified to include
information about the type of the property itself, which was previously
missing.

Fixes: QTBUG-102088
Pick-to: 6.3 6.2
Change-Id: I7007614e8db21164e9525c160c6ea124218c5bb3
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>