QDoc: Show all inherited members in the all members pages
QDoc produces a reference page containing all members of a class both
inherited and "owned".

The list was incomplete missing some of the inherited members.
For example, in the all members page for `QAbstractItemView`, the
`update()` slot from `QWidget` was missing.

QDoc generates the list of members for those pages in `sections.cpp`,
aggregating the various members in a multimap trough repeated calls to
`Section::insert`.

In the `insert` method, the code that took care of collecting all the
members, so that the all members page can later be generated,
specifically avoided collecting members whose name collided with the
name of some other member.

The name that is used is based on the one produced by
`Section::sortName` which doesn't necessarily produce unique names.

Indeed, in the `QWidget::update()` case, the produced name collided with
the owned `QAbstractItemView::update(const QModelIndex&)` one, such that
it was not added to the relevant list of members for the all members
page.

To fix the issue the check for collision was removed in favor of always
collecting all members.

It is unclear if there was any historic purpose for the check and why it
would be used along with a multimap.
The removal of it does not seem to produce any unintended side effect on
the current documentation, such that the only noticeable effect is that
of adding the missing members to the all members pages.

The output files for `tst_generatedOutput` were regenerated to take into
account the changes.

Fixes: QTBUG-102342
Pick-to: 6.3 6.2
Change-Id: I8f63b9d63e8be2cd2d65e62f91a0b6666935aada
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>