doc: Add some bits about using the \fn command for hidden friends
The Qt project uses hidden friends quite often as a result of the
changes started in QTBUG-87973.

QDoc is able to document hidden friends but requires the relevant
comment-block to refer to the friend as a member of the enclosing class.

This syntax doesn't mimic what could be an intuitive way, from a C++
point of view, to refer to the friend and has proved to be confusing in
the past (QTBUG-101649).

As we don't currently plan on relaxing this requirement, some bits of
documentation was added to the QDoc manual, under the `\fn` command
section, to warn about documenting hidden friends.

Change-Id: Iabb7f991804cb5c5cae98284c79b98104d04e8b4
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>