qdoc: Use QML-specific code marker when parsing a QML topic
The \code command is used widely to document also QML code snippets,
however, QDoc failed to detect the code as QML because the default
code marker is the CppCodeMarker. The consequence was that QDoc
failed to detect and auto-link to QML types in code snippets.

The default marker to use depends on the 'language' configuration
variable and pure QML documentation projects can set it to QML
instead. Many projects consist of both C++ and QML module docs
however, and using a single project-wide setting is less than
optimal.

As a compromise, use the QML marker for \code snippets appearing
within QML-specific topics as the default code marker. Ensure
that the code is detected as QML - this prevents unexpected
warnings for parsing failures typically seen with the \qml
command.

Note that the downside of the chosen solution is that QML topic
command must appear before the \code commands for the automatic
marker selection to take place. This can't be helped as all
commands are parsed sequentially and we don't know the topic
beforehand. Typically this is not a problem as the convention is
to start the documentation comment with a topic command, even
though QDoc doesn't strictly require it.

Task-number: QTBUG-103791
Change-Id: I87f2549c7c258a15fc02f64f65ef7fa0b28b5a68
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>