qdoc: Down-prioritize linking to section titles
It was too easy to inadvertently divert links generated by QDoc by
simply introducing a \section with a title that duplicates an existing
link target in the documentation. This happened quite frequently, for
example, by having names of Qt modules as section titles.

Treat section titles as a fallback when resolving links; internally
within a single Tree, section titles were already the lowest
priority candidates; extend this to apply to the search across the
entire forest of trees.

This has a slight performance impact for modules that pull in a lot of
dependencies (e.g. qtdoc), but it's arguably worth it.

Pick-to: 6.4
Fixes: QTBUG-104137
Change-Id: I2904c7023690369ea1c707791b25a9b0cb0361fd
Reviewed-by: Luca Di Sera <luca.disera@qt.io>
Reviewed-by: Kai Koehne <kai.koehne@qt.io>