QDoc: Ensure ordering for the members section of a since list
When QDoc generates a list from a `\sincelist` command, it generates,
between the others, a list of newly added member functions, qml methods
and qml properties.

The list groups and orders related members based on their containing
element, such as a class or qml item, and shows them based on this
grouping.

Internally, this ordering is ensured by the usage of QMaps, whose
iteration is deterministic based on its keys.

Nonetheless, the map that was used for the ordering and grouping of
related members used a pointer to the containing element itself, as a
`Node`, such that the ordering depended on the addresses of the nodes
themselves and not on any of their distinguishing elements.

Furthermore, this ordering may have changed from generation to the
other, as the address of each element may not be stable.

To avoid the issue and to improve the presentation of the resulting list
by using an obvious ordering based on the containing element's name,
which is what the user is showed in the final product; the internal
map was modified to use the `fullName`, inasmuch as it is a totally
ordered unique identifier, of the relevant nodes as its keys.

Do note that this is a slight pessimization as it requires multiple
calls to `fullName` and various comparisons of strings instead of some
simple address comparisons.
It would have been possible to avoid some of the newly added overhead
but it would have required some deeper changes to related parts of the
code and was thus ignored, with the consideration that the pessimization
is expected to be unnoticeable.

A type alias in `generator.h`, `ParentMaps`, that was used only in the
code relevant to this issue was removed as it was unnecessary and added
some indirection to the reading of the code at its usage point.

Fixes: QTBUG-104216
Pick-to: 6.4
Change-Id: I4b8a9e374337d7c72a2d6e50fa4f6c5874a9fc5b
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>