qdoc: Always consider section titles as the last resort for linking
This commit amends a8cb221c. There was one more situation where a
section title could mask a more high-priority link target in the same
project.

This happened because in Tree::findNodeForTarget(), we search the
target map first as it's a relatively fast operation. If the target
is found but refers to a section title, delay returning the node as
there may be higher-priority targets in the same tree.

Pick-to: 6.4
Fixes: QTBUG-104369
Change-Id: Ifecf94fb09b3484dac6d3329b8624d26b3e26a2c
Reviewed-by: Luca Di Sera <luca.disera@qt.io>