QDoc: Add reference to groups in class/qmltype pages
QDoc allows to group related elements through the use of the "\ingroup"
command.
Furthermore, it allows the generation of a page for a certain group,
through the "\group" command, to provide information about the way that
the elements are related and a list with links to element that are
member of the given group.

While there is an obvious connection from the page of a group to its
elements, there currently is no such inverse connection, reducing the
discoverability of those pages for the user of the documentation.

To reduce this issue, a line providing links to the available pages of groups
that a class or qmltype element is a child of is now added to respective
element pages.

QDoc allows to include an element into some group without generating a
page for the group. In particular, this allows the addition of elements
that are defined in a certain project in groups whose page is defined in
a different project.

To support the cross-project linking that those cases require, some
changes were made to the way in which QDoc finds source-of-truth
information for a group node.

QDoc stores a group every time an "\ingroup" command is found. When a
"\group" command is found, the stored group is enhanced with the
provided information.

If the "\group" command for a certain group is not in the same project
as some of its "\ingroup" declaration, the information about the group
must be read from an index file.

When this is done, assuming that the correct dependency are declared in
the configuration file, QDoc ends up with multiple representation for
the same group, only one of which, assuming that no repeated declaration
for the same group exists, which is assumed to be how group are
implicitly supposed to work, albeit this might not be enforced, contains
complete information, which are required for linking.

Generally, if the group is defined in a depended upon module, the directly
accessible group representation, the one on the primary tree, will not
contain complete information, inhibiting linking.

QDoc further processes those group representation, generally for the one
on the primary tree, merging them with the information from the other
representations of the same group that were read from the index files, to
ensure that a correct list of members, containing the members that are
declared in other projects, is generated.

To ensure that correct linking information are available as long as a
group definition exists, either in the current project or one of its
dependencies, this merging procedure was modified to merge any
information that is required for linking purposes, so that it is
available in the directly accessible representation of the group.

As this merging procedure is performed when needed, such that it must be
performed each time a group is accessed for information, as it might not
have been merged before, a possibly high overhead might be generated
when the group is accessed many times.

For this feature in particular, access to the group is required for each
class and qmltype pages, increasing the amount of expected merges.

To lessen the impact that this would have on performance, a hackish
caching was added to the merging procedure, to avoid merging more than
once.

The merging procedure is only called during the generation phase and is
only meaningful when some representation of a group was generated from
an index file.

As the index files are always read much before the generation phase and
the primary group representation itself is not expected to change during
the same phase, the result of a merge should be the same at any point in
the generation, such that the cache should never invalidate and no
precautions for cache invalidation were provided.

The regression files for `tst_generatedOutput` were regenerated to take
the changes into account.

Fixes: QTBUG-104237
Change-Id: Ia2332a7b91f7d16c16899c2de9920175bc57f177
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
Reviewed-by: Kai Koehne <kai.koehne@qt.io>