QDoc: Improve failure report for tst_QDoc
For many tests in QDoc we make use of randomly generated input to
validate our assumptions over a domain.

Generally, if a test fails, the generated values will be shown in the
failure report as long as they appear "naked" in the failing assertion.

Indeed, most of the tests should strive to write each assertion as a
form of comparison to the generated value.

Nonetheless, this is not always possible as some tests may require more
complex assertions or may have consumed the value earlier to produce the
actually tested value, the representation of which may render the
generated element opaque.

To improve the failure report so that spurious edge cases can be
correctly debugged, a CAPTURE directive was included for the generated
elements in those tests that are complex enough to require the adequate
generative context but are unable to provide it through an assertion
context.

The addition of the CAPTURE directive will ensure that the generated
value will be represented in the respective failure report.

For example, if we have some `T a = GENERATE(Phi)`, where T is some type
and Phi is an expression whose result is a generator of Ts, a CAPTURE(a)
directive will show a line similar to "a := <<Phi_n>>" where Phi_n is the
value of T that was generated during the failure and <<Phi_n>> is a
printable representation of Phi_n.

This ensures that we can reproduce and debug the generation of test
cases that have shown failures by temporarily modifying the generation
clause. For example, `T a = GENERATE((Phi_n), Phi)`, where (Phi_n) is an
expression that produces a value that is structurally equivalent
to Phi_n, which will always generate the known failure value as its
first value.

Change-Id: Ide207068c6d5d36a81a53bfb2ee6357eec76db5d
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>