QDoc: Avoid generating paths that are not cleanable in tst_QDoc
"tst_QDoc" tests "DirectoryPath", which, as part of its definition,
requires testing the effect of permission changes on a path representing
a directory (on *nix systems only).

As with all testing that deal with actual elements on the filesystem, we
constrain our tests to a temporary directory that is later cleaned up,
such as to avoid leaving garbage after a test run.

To do this, we always restore all possible permissions on the path after
the test has finished its run.

Nonetheless, this may fail, depending on the changed permissions, for
paths ending in a relative component.

On *nix systems "." and ".." are actual nodes that are under each
directory.
As elements of the directory, they need certain permissions for certain
operations to be performed on them.

When the permissions of a relative-component-ending path are changed,
the one of the containing or parent directory may change too, so that it
is not possible to reset the permission on the path itself, which would
require access to the "." or ".." node.

This in turn would disallow the cleanup of the temporary directories,
either leaving garbage on the filesystem or, in our case, failing the
test.

To avoid this issue, relative directory paths for tests related to
permissions in "catch_directorypath.cpp" will now filter out paths that
end in a relative component.

Do note that this is a test-implementation detail and does not affect
QDoc itself. Indeed, the test was still validly asserting the required
behavior.

With this change, we slightly neuter the execution permission test for
"DirectoryPath", as a relative-component-ending path was one of the main
cases for it.

This should be changed later on in case a valid structure for the test
that doesn't incur in the same limitation is found, so that the validity
of the test can be restored to full again.

Change-Id: Icfe9f0d04ac44fabe1ad3031e4844b4a7ab93fac
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>