qdoc: Make types inherit \since statement from their module docs
When introducing a new module (C++ or QML) documentation, the \module or
\qmlmodule topic should be marked with a \since command. Types belonging
to the module will then inherit the \since by default, unless explicitly
specified in the type documentation.

Unify the interface of a similar Tree::resolveUsingClauses() method to
explicitly require a reference to an Aggregate for the recursive
operation.

Fixes: QTBUG-104613
Change-Id: I48afe82e55a889dab43d8c59882c52d44b8bb69a
Reviewed-by: Luca Di Sera <luca.disera@qt.io>