Use registerRef to ensure xml:id validity and unicity
Before this change, invalid xml:id were generated:
- duplicates (several tags having the same xml:id -- not several ones pointing at the same ID, which is fine)
- invalid characters (mostly spaces)
- IDs not starting with a letter or an underscore (in particular, digits are not allowed in XML)
- colons in IDs. Only few occurrences of this problem:
  - https://doc.qt.io/qt-6/qglobalstatic.html#operator-QGlobalStatic::Type--2a
  - https://doc.qt.io/qt-6/qflags.html#operator-QFlags::Int

A flag controls the changes that are applied on IDs, so that there should be no changes for HTML IDs (very lax).

Furthermore, stop using registerRef when it is not needed and clear the set of generated xml:id for each new document.

Pick-to: 6.4
Change-Id: I53b08ae884a4e360ff4287fd486e24fb17b3deb2
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>