In an annotated list, wrap the child nodes in a section if needed
The need may arise when there is already a section
in the document: there cannot be text outside and
after sections (before them is fine).

This variable should only be reset when creating a
new document, because it is not yet known if it has
sections or not. Within the same XML output file,
the value should be kept to true once a section is met.

Change-Id: I297df8497b56230b960b3ad70fa9f11364978c1e
Reviewed-by: Paul Wicking <paul.wicking@qt.io>