Refactor to use the startSection/endSection helpers
On the way, introduce a new overload of startSection that doesn't take an ID (used several times, allows to add a comment about this) and one that takes a node (to map recent changes in writeXmlId).

This change also includes some sections that were previously missing and lead to DocBook-validity issues.

Pick-to: 6.4
Change-Id: I515f3e31298462997c210a8063db2caaae80dd1d
Reviewed-by: Paul Wicking <paul.wicking@qt.io>