lrelease: Fix reporting of duplicates
This reverts commit 3af741cc180445cc487af6853575651ba204c4c1 and fixes
QTBUG-86507 differently.

The fix for QTBUG-86507 modified the algorithm to not remove elements
from m_messages while looping over it.  However, that changes the
indices that are returned by resolveDuplicates.

Bring back the old algorithm and change the TranslateMessage*Ptr classes
to hold indices to elements of m_messages instead of pointers.

Pick-to: 6.2 6.3 6.4
Task-number: QTBUG-86507
Change-Id: Ib395fa48da4e0d688ac7f2b766134c7711412b1b
Reviewed-by: hjk <hjk@qt.io>