qdoc: Store BINDABLE property members into PropertyNode
QDoc already recognized the BINDABLE attribute of Q_PROPERTY
but did not store the name of the associated member.

Store it in the list of a PropertyNode's access functions.
This information will be useful later when deciding on
whether to generate a warning for missing member functions;
functions that are associated with properties can be left
undocumented if the property itself is documented.

Add a static helper function PropertyNode that returns the
names of property access function roles. This is used when
writing the access functions into an .index file.

Convert access function-related enums to enum classes.

Change-Id: Idd59622a965895883acf3ae297297ba3f3cca20e
Reviewed-by: Luca Di Sera <luca.disera@qt.io>